Conflicts:
	bridge.c
	main.c
	message.c
	types.h

+++ bridge.c
@@ -43,7 +43,7 @@ int8_t bridge_releaseMemoryBlock(void *blk) {
     int8_t status = SUCCESS;
     PCB *firstBlocked = NULL;
 
-    status = k_releaseMemoryBlock(&gMem, blk, gProcInfo.currentProcess->pid);
+    status = k_releaseMemoryBlock(&gMem, (uint32_t)blk, gProcInfo.currentProcess->pid);
     if (status != SUCCESS) {
         return status;
     }
